+++ tests/testthat/test-exoplanets.R
@@ -1,10 +1,3 @@
-# remove all messages from `exoplanets` for a clean test log
-quiet <- function(x) {
-  sink(tempfile())
-  on.exit(sink())
-  suppressMessages(invisible(force(x)))
-}
-
 test_that("exoplanets is memoised", {
   expect_true(memoise::is.memoised(exoplanets))
 })
@@ -48,31 +41,31 @@ test_that("parse_url works", {
 
 with_mock_dir("exoplanets-default", {
   test_that("exoplanets default works", {
-    expect_s3_class(quiet(exoplanets("k2names")), "data.frame")
+    expect_s3_class(exoplanets("k2names", quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-csv", {
   test_that("exoplanets csv works", {
-    expect_s3_class(quiet(exoplanets("k2names", format = "csv")), "data.frame")
+    expect_s3_class(exoplanets("k2names", format = "csv", quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-tsv", {
   test_that("exoplanets tsv works", {
-    expect_s3_class(quiet(exoplanets("k2names", format = "tsv")), "data.frame")
+    expect_s3_class(exoplanets("k2names", format = "tsv", quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-json", {
   test_that("exoplanets json works", {
-    expect_true(inherits(quiet(exoplanets("k2names", format = "json")), "list"))
+    expect_true(inherits(exoplanets("k2names", format = "json", quiet = TRUE), "list"))
   })
 })
 
 with_mock_dir("exoplanets-default-select", {
   test_that("exoplanets default works when selecting", {
-    r <- quiet(exoplanets("k2names", c("epic_id", "k2_name")))
+    r <- exoplanets("k2names", c("epic_id", "k2_name"), quiet = TRUE)
     expect_s3_class(r, "data.frame")
     expect_true(all(c("epic_id", "k2_name") %in% names(r)))
   })
@@ -80,7 +73,7 @@ with_mock_dir("exoplanets-default-select", {
 
 with_mock_dir("exoplanets-csv-select", {
   test_that("exoplanets csv works when selecting", {
-    r <- quiet(exoplanets("k2names", c("epic_id", "k2_name"), format = "csv"))
+    r <- exoplanets("k2names", c("epic_id", "k2_name"), format = "csv", quiet = TRUE)
     expect_s3_class(r, "data.frame")
     expect_true(all(c("epic_id", "k2_name") %in% names(r)))
   })
@@ -88,7 +81,7 @@ with_mock_dir("exoplanets-csv-select", {
 
 with_mock_dir("exoplanets-tsv-select", {
   test_that("exoplanets tsv works when selecting", {
-    r <- quiet(exoplanets("k2names", c("epic_id", "k2_name"), format = "tsv"))
+    r <- exoplanets("k2names", c("epic_id", "k2_name"), format = "tsv", quiet = TRUE)
     expect_s3_class(r, "data.frame")
     expect_true(all(c("epic_id", "k2_name") %in% names(r)))
   })
@@ -96,7 +89,7 @@ with_mock_dir("exoplanets-tsv-select", {
 
 with_mock_dir("exoplanets-json-select", {
   test_that("exoplanets json works when selecting", {
-    r <- quiet(exoplanets("k2names", c("epic_id", "k2_name"), format = "json"))
+    r <- exoplanets("k2names", c("epic_id", "k2_name"), format = "json", quiet = TRUE)
     expect_true(inherits(r, "list"))
     expect_true(all(c("epic_id", "k2_name") %in% unique(unlist(lapply(r, names)))))
   })
@@ -105,48 +98,48 @@ with_mock_dir("exoplanets-json-select", {
 with_mock_dir("exoplanets-table-ps", {
   test_that("ps table works", {
     column <- "pl_name"
-    expect_s3_class(quiet(exoplanets("ps", column)), "data.frame")
+    expect_s3_class(exoplanets("ps", column, quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-table-pscomppars", {
   test_that("pscomppars table works", {
     column <- "pl_name"
-    expect_s3_class(quiet(exoplanets("pscomppars", column)), "data.frame")
+    expect_s3_class(exoplanets("pscomppars", column, quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-table-keplernames", {
   test_that("keplernames table works", {
     column <- "pl_name"
-    expect_s3_class(quiet(exoplanets("keplernames", column)), "data.frame")
+    expect_s3_class(exoplanets("keplernames", column, quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-table-k2names", {
   test_that("k2names table works", {
     column <- "pl_name"
-    expect_s3_class(quiet(exoplanets("k2names", column)), "data.frame")
+    expect_s3_class(exoplanets("k2names", column, quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-table-emissionspec", {
   test_that("emissionspec table works", {
     column <- "centralwavelng"
-    expect_s3_class(quiet(exoplanets("emissionspec", column)), "data.frame")
+    expect_s3_class(exoplanets("emissionspec", column, quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-table-transitspec", {
   test_that("transitspec table works", {
     column <- "centralwavelng"
-    expect_s3_class(quiet(exoplanets("transitspec", column)), "data.frame")
+    expect_s3_class(exoplanets("transitspec", column, quiet = TRUE), "data.frame")
   })
 })
 
 with_mock_dir("exoplanets-example-1", {
   test_that("example 1 works", {
-    r <- quiet(exoplanets("ps", c("pl_name", "discoverymethod")))
+    r <- exoplanets("ps", c("pl_name", "discoverymethod"), quiet = TRUE)
     expect_s3_class(r, "data.frame")
     expect_true(all(c("pl_name", "discoverymethod") %in% names(r)))
   })
@@ -154,7 +147,7 @@ with_mock_dir("exoplanets-example-1", {
 
 with_mock_dir("exoplanets-example-2", {
   test_that("example 2 works", {
-    r <- quiet(exoplanets("ps", c("pl_name", "discoverymethod"), format = "json"))
+    r <- exoplanets("ps", c("pl_name", "discoverymethod"), format = "json", quiet = TRUE)
     expect_true(inherits(r, "list"))
     expect_true(all(c("pl_name", "discoverymethod") %in% names(r[[1]])))
   })
@@ -162,7 +155,13 @@ with_mock_dir("exoplanets-example-2", {
 
 with_mock_dir("exoplanets-example-3", {
   test_that("example 3 works", {
-    r <- quiet(exoplanets("k2names", progress = FALSE))
+    r <- exoplanets("k2names", progress = FALSE, quiet = TRUE)
     expect_s3_class(r, "data.frame")
+  })
+})
+
+with_mock_dir("exoplanets-quiet", {
+  test_that("quiet works", {
+    expect_output(exoplanets("k2names", quiet = TRUE), NA)
   })
 })